Move signal filter logic to class
